requirements.txt
Django==4.2.3
asgiref==3.8.1
sqlparse==0.5.0
daphne==4.1.2
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-allauth==0.62.1
python-dotenv==1.0.1
mysqlclient==2.2.4
channels==4.1.0
django-crispy-forms==2.1
crispy-bootstrap4==2024.1
python-docx